empty

+++ YAD/YADtools.h
@@ -97,7 +97,7 @@ public:
     ~Truth() { file-> Close(); }
     
     size_t GetEntries() { return tree->GetEntries(); }
-    void GetEntry(int i) { tree->GetEntry(i); }
+    void GetEntry(size_t i) { tree->GetEntry(i); }
 };
 
 class Reco {
@@ -214,7 +214,7 @@ public:
     ~Reco() { file-> Close(); }
     
     size_t GetEntries() { return tree->GetEntries(); }
-    void GetEntry(int i) { tree->GetEntry(i); }
+    void GetEntry(size_t i) { tree->GetEntry(i); }
     
 };
